+++ src/Layouts/Lock/style.scss
@@ -1,5 +1,5 @@
 $phone-heght: 99vh;
-$screen-heght: $phone-heght * 0.927;
+$screen-heght: $phone-heght * 0.94;
 $screen-width: ($phone-heght/19.5) * 9;
 $screen-radius: $phone-heght/18.8;
